--- front/components/assistant/conversation/ValidationRequirementsProvider.tsx
@@ -1,457 +1,0 @@
-import type { MultiPageDialogPage } from "@dust-tt/sparkle";
-import {
-  ActionPieChartIcon,
-  Checkbox,
-  CodeBlock,
-  CollapsibleComponent,
-  Icon,
-  Label,
-  Spinner,
-} from "@dust-tt/sparkle";
-import { MultiPageDialog, MultiPageDialogContent } from "@dust-tt/sparkle";
-import type { ReactNode } from "react";
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
-
-import { useNavigationLock } from "@app/components/assistant_builder/useNavigationLock";
-import type { MCPValidationOutputType } from "@app/lib/actions/constants";
-import { getAvatarFromIcon } from "@app/lib/actions/mcp_icons";
-import { useBlockedActions } from "@app/lib/swr/blocked_actions";
-import type {
-  ConversationWithoutContentType,
-  LightWorkspaceType,
-  MCPActionValidationRequest,
-} from "@app/types";
-import { asDisplayName, pluralize } from "@app/types";
-
-// Types for validation requirements
-interface ValidationRequirement {
-  id: string;
-  actionId: string;
-  conversationId: string;
-  messageId: string;
-  toolName: string;
-  mcpServerName: string;
-  agentName: string;
-  inputs: Record<string, unknown>;
-  stake?: string;
-  icon?: string;
-}
-
-interface ValidationRequirementsContextType {
-  showValidationDialog: (
-    validationRequest?: MCPActionValidationRequest
-  ) => void;
-  hasPendingValidations: boolean;
-  totalPendingValidations: number;
-}
-
-const ValidationRequirementsContext =
-  createContext<ValidationRequirementsContextType | null>(null);
-
-export function useValidationRequirementsContext() {
-  const context = useContext(ValidationRequirementsContext);
-  if (!context) {
-    throw new Error(
-      "useValidationRequirementsContext must be used within ValidationRequirementsProvider"
-    );
-  }
-  return context;
-}
-
-interface ValidationRequirementsProviderProps {
-  owner: LightWorkspaceType;
-  conversation: ConversationWithoutContentType | null;
-  children: ReactNode;
-}
-
-export function ValidationRequirementsProvider({
-  owner,
-  conversation,
-  children,
-}: ValidationRequirementsProviderProps) {
-  const { blockedActions, mutate: mutateBlockedActions } = useBlockedActions({
-    conversationId: conversation?.sId || null,
-    workspaceId: owner.sId,
-  });
-
-  // Filter blocked actions to only get validation required ones
-  const validationRequirements = useMemo(() => {
-    return blockedActions
-      .filter((action) => action.status === "blocked_validation_required")
-      .map((action) => ({
-        id: `${action.actionId}`,
-        actionId: action.actionId,
-        conversationId: action.conversationId,
-        messageId: action.messageId,
-        toolName: action.metadata.toolName,
-        mcpServerName: action.metadata.mcpServerName,
-        agentName: action.metadata.agentName,
-        inputs: action.inputs,
-        stake: action.stake,
-        icon: action.metadata.icon,
-      }));
-  }, [blockedActions]);
-
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const [currentPageId, setCurrentPageId] = useState<string>("overview");
-  const [isProcessing, setIsProcessing] = useState(false);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
-  const [neverAskAgain, setNeverAskAgain] = useState(false);
-  const [completedValidations, setCompletedValidations] = useState<Set<string>>(
-    new Set()
-  );
-
-  useNavigationLock(isDialogOpen);
-
-  // Poll for blocked actions every 2 seconds when conversation is active
-  useEffect(() => {
-    if (!conversation?.sId) {
-      return;
-    }
-
-    const interval = setInterval(() => {
-      void mutateBlockedActions();
-    }, 2000);
-
-    return () => clearInterval(interval);
-  }, [conversation?.sId, mutateBlockedActions]);
-
-  // Automatically open dialog when validation requirements become available
-  useEffect(() => {
-    console.log(
-      "ValidationRequirementsProvider: validationRequirements changed",
-      {
-        length: validationRequirements.length,
-        isDialogOpen,
-        requirements: validationRequirements,
-      }
-    );
-
-    if (validationRequirements.length > 0 && !isDialogOpen) {
-      console.log(
-        "ValidationRequirementsProvider: Opening dialog for",
-        validationRequirements.length,
-        "validations"
-      );
-      // Skip overview page if there's only one validation
-      if (validationRequirements.length === 1) {
-        setCurrentPageId(`validation-${validationRequirements[0].id}`);
-      } else {
-        setCurrentPageId("overview");
-      }
-      setIsDialogOpen(true);
-      setCompletedValidations(new Set());
-      setErrorMessage(null);
-    }
-  }, [validationRequirements]);
-
-  const sendValidation = useCallback(
-    async (
-      requirement: ValidationRequirement,
-      status: MCPValidationOutputType
-    ) => {
-      let approved = status;
-      if (status === "approved" && neverAskAgain) {
-        approved = "always_approved";
-      }
-
-      setErrorMessage(null);
-      setIsProcessing(true);
-
-      try {
-        const response = await fetch(
-          `/api/w/${owner.sId}/assistant/conversations/${requirement.conversationId}/messages/${requirement.messageId}/validate-action`,
-          {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify({
-              actionId: requirement.actionId,
-              approved,
-            }),
-          }
-        );
-
-        if (!response.ok) {
-          throw new Error("Failed to assess action approval");
-        }
-
-        setCompletedValidations((prev) => new Set([...prev, requirement.id]));
-        setNeverAskAgain(false);
-
-        // Revalidate blocked actions to update the UI
-        await mutateBlockedActions();
-
-        return true;
-      } catch (error) {
-        setErrorMessage("Failed to assess action approval. Please try again.");
-        return false;
-      } finally {
-        setIsProcessing(false);
-      }
-    },
-    [owner.sId, neverAskAgain, mutateBlockedActions]
-  );
-
-  const handleValidationSubmit = useCallback(
-    async (
-      requirement: ValidationRequirement,
-      approved: MCPValidationOutputType
-    ) => {
-      const success = await sendValidation(requirement, approved);
-
-      if (!success) {
-        return; // Don't proceed if validation failed
-      }
-
-      // Find the next uncompleted validation in order
-      const remainingValidations = validationRequirements.filter(
-        (req) => !completedValidations.has(req.id) && req.id !== requirement.id
-      );
-
-      if (remainingValidations.length > 0) {
-        const nextRequirement = remainingValidations[0];
-        setCurrentPageId(`validation-${nextRequirement.id}`);
-      } else {
-        // All validations completed, close dialog
-        setIsDialogOpen(false);
-        setCurrentPageId("overview");
-        setCompletedValidations(new Set());
-      }
-    },
-    [sendValidation, validationRequirements, completedValidations]
-  );
-
-  const showValidationDialog = useCallback(() => {
-    if (validationRequirements.length > 0) {
-      setIsDialogOpen(true);
-      // Skip overview page if there's only one validation
-      if (validationRequirements.length === 1) {
-        setCurrentPageId(`validation-${validationRequirements[0].id}`);
-      } else {
-        setCurrentPageId("overview");
-      }
-      setCompletedValidations(new Set());
-      setErrorMessage(null);
-    }
-  }, [validationRequirements]);
-
-  const hasPendingValidations = validationRequirements.length > 0;
-  const totalPendingValidations = validationRequirements.length;
-
-  // Create dialog pages
-  const dialogPages: MultiPageDialogPage[] = useMemo(() => {
-    if (validationRequirements.length === 0) {
-      return [];
-    }
-
-    const pages: MultiPageDialogPage[] = [
-      // Overview page
-      {
-        id: "overview",
-        title: "Tool Validation Required",
-        description: `${validationRequirements.length} tool${pluralize(validationRequirements.length)} need approval`,
-        content: (
-          <div className="flex flex-col gap-4">
-            <div className="text-sm text-muted-foreground">
-              The following tools require your approval before they can be used:
-            </div>
-            <div className="space-y-3">
-              {validationRequirements.map((req, index) => (
-                <div
-                  key={req.id}
-                  className="flex items-center justify-between rounded-lg border p-3"
-                >
-                  <div className="flex items-center gap-3">
-                    <div className="flex h-8 w-8 items-center justify-center rounded-full bg-muted">
-                      {req.icon ? (
-                        getAvatarFromIcon(req.icon)
-                      ) : (
-                        <Icon visual={ActionPieChartIcon} size="sm" />
-                      )}
-                    </div>
-                    <div>
-                      <div className="font-medium">
-                        {asDisplayName(req.toolName)}
-                      </div>
-                      <div className="text-sm text-muted-foreground">
-                        from {asDisplayName(req.mcpServerName)}
-                      </div>
-                    </div>
-                  </div>
-                  <div className="text-sm text-muted-foreground">
-                    {index + 1} of {validationRequirements.length}
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        ),
-      },
-    ];
-
-    // Add individual validation pages in the same order as blockedActions
-    validationRequirements.forEach((req) => {
-      pages.push({
-        id: `validation-${req.id}`,
-        title: `Validate ${asDisplayName(req.toolName)}`,
-        description: `Allow @${req.agentName} to use this tool`,
-        content: (
-          <div className="flex flex-col gap-4">
-            <div>
-              Allow <b>@{req.agentName}</b> to use the tool{" "}
-              <b>{asDisplayName(req.toolName)}</b> from{" "}
-              <b>{asDisplayName(req.mcpServerName)}</b>?
-            </div>
-
-            {req.inputs && Object.keys(req.inputs).length > 0 && (
-              <CollapsibleComponent
-                triggerChildren={
-                  <span className="font-medium text-muted-foreground">
-                    Details
-                  </span>
-                }
-                contentChildren={
-                  <div>
-                    <div className="max-h-80 overflow-auto rounded-lg bg-muted">
-                      <CodeBlock
-                        wrapLongLines
-                        className="language-json overflow-y-auto"
-                      >
-                        {JSON.stringify(req.inputs, null, 2)}
-                      </CodeBlock>
-                    </div>
-                  </div>
-                }
-              />
-            )}
-
-            {errorMessage && (
-              <div className="text-sm font-medium text-warning-800">
-                {errorMessage}
-              </div>
-            )}
-
-            {req.stake === "low" && (
-              <div className="mt-5">
-                <Label className="copy-sm flex w-fit cursor-pointer flex-row items-center gap-2 py-2 pr-2 font-normal">
-                  <Checkbox
-                    checked={neverAskAgain}
-                    onCheckedChange={(check) => {
-                      setNeverAskAgain(!!check);
-                    }}
-                  />
-                  <span>Always allow this tool</span>
-                </Label>
-              </div>
-            )}
-          </div>
-        ),
-      });
-    });
-
-    return pages;
-  }, [validationRequirements, errorMessage, neverAskAgain]);
-
-  // Get current page to determine which buttons to show
-  const isOverviewPage = currentPageId === "overview";
-  const currentValidationRequirement = isOverviewPage
-    ? null
-    : validationRequirements.find(
-        (req) => currentPageId === `validation-${req.id}`
-      );
-
-  // Configure buttons based on current page
-  const getDialogButtons = () => {
-    if (isOverviewPage) {
-      return {
-        leftButton: {
-          label: "Cancel",
-          variant: "outline" as const,
-          onClick: () => setIsDialogOpen(false),
-        },
-        rightButton: {
-          label: "Start Review",
-          variant: "highlight" as const,
-          onClick: () => {
-            const firstRequirement = validationRequirements[0];
-            if (firstRequirement) {
-              setCurrentPageId(`validation-${firstRequirement.id}`);
-            }
-          },
-        },
-      };
-    }
-
-    if (currentValidationRequirement) {
-      return {
-        leftButton: {
-          label: "Decline",
-          variant: "outline" as const,
-          onClick: () =>
-            handleValidationSubmit(currentValidationRequirement, "rejected"),
-          disabled: isProcessing,
-          children: isProcessing ? (
-            <div className="flex items-center">
-              <span className="mr-2">Declining</span>
-              <Spinner size="xs" variant="dark" />
-            </div>
-          ) : undefined,
-        },
-        rightButton: {
-          label: "Allow",
-          variant: "highlight" as const,
-          onClick: () =>
-            handleValidationSubmit(currentValidationRequirement, "approved"),
-          disabled: isProcessing,
-          autoFocus: true,
-          children: isProcessing ? (
-            <div className="flex items-center">
-              <span className="mr-2">Approving</span>
-              <Spinner size="xs" variant="light" />
-            </div>
-          ) : undefined,
-        },
-      };
-    }
-
-    return {};
-  };
-
-  const dialogButtons = getDialogButtons();
-
-  return (
-    <ValidationRequirementsContext.Provider
-      value={{
-        showValidationDialog,
-        hasPendingValidations,
-        totalPendingValidations,
-      }}
-    >
-      {children}
-
-      {dialogPages.length > 0 && (
-        <MultiPageDialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
-          <MultiPageDialogContent
-            pages={dialogPages}
-            currentPageId={currentPageId}
-            onPageChange={setCurrentPageId}
-            size="lg"
-            showNavigation={validationRequirements.length > 1}
-            showHeaderNavigation={validationRequirements.length > 1}
-            addFooterSeparator={true}
-            {...dialogButtons}
-          />
-        </MultiPageDialog>
-      )}
-    </ValidationRequirementsContext.Provider>
-  );
-}
