+++ front/components/assistant_builder/AssistantBuilderPreviewDrawer.tsx
@@ -22,13 +22,13 @@ import assert from "assert";
 import { useContext, useEffect, useRef, useState } from "react";
 
 import { AssistantDetailsPerformance } from "@app/components/assistant/AssistantDetailsPerformance";
+import { ActionValidationProvider } from "@app/components/assistant/conversation/ActionValidationProvider";
 import ConversationSidePanelContent from "@app/components/assistant/conversation/ConversationSidePanelContent";
 import { useConversationSidePanelContext } from "@app/components/assistant/conversation/ConversationSidePanelContext";
 import { ConversationsNavigationProvider } from "@app/components/assistant/conversation/ConversationsNavigationProvider";
 import ConversationViewer from "@app/components/assistant/conversation/ConversationViewer";
 import { GenerationContextProvider } from "@app/components/assistant/conversation/GenerationContextProvider";
 import { AssistantInputBar } from "@app/components/assistant/conversation/input_bar/InputBar";
-import { ValidationRequirementsProvider } from "@app/components/assistant/conversation/ValidationRequirementsProvider";
 import { useAssistantBuilderContext } from "@app/components/assistant_builder/contexts/AssistantBuilderContexts";
 import {
   usePreviewAssistant,
@@ -213,7 +213,7 @@ export default function AssistantBuilderRightPanel({
               </div>
             ) : (
               <ConversationsNavigationProvider>
-                <ValidationRequirementsProvider
+                <ActionValidationProvider
                   owner={owner}
                   conversation={conversation}
                 >
@@ -273,7 +273,7 @@ export default function AssistantBuilderRightPanel({
                       currentPanel={currentPanel}
                     />
                   </GenerationContextProvider>
-                </ValidationRequirementsProvider>
+                </ActionValidationProvider>
               </ConversationsNavigationProvider>
             )}
           </div>
